+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added:
 
+- Installer for `yazi` and my `yazi` cfg files
+    - Wired in the `yazi` installer into install and check process
 - New depdencies to fix build errors in `Trixie`
 - Debian version package awareness to ``00-depdencies.sh`
     - Stops failed install errors as I tried to install all versions of some pkgs
@@ -17,6 +19,7 @@
 
 ### Fixed:
 
+- Not all scripts used `#!/usr/bin/env bash`
 - Some scritps not executable
 - LUA pkg name in trixie, fixinng installation failure errors
 - hardcode username is `hyprtoolkit`